feat(chat): refactor history tree file structure
