+++ fai-rag-app/fai-sveltekit-frontend/src/lib/components/Menu/Chat/HistoryTree/HistoryTree.stories.svelte
@@ -1,6 +1,6 @@
 <script module>
   import { defineMeta } from '@storybook/addon-svelte-csf'
-  import HistoryTree from '$lib/components/Menu/Chat/HistoryTree.svelte'
+  import HistoryTree from '$lib/components/Menu/Chat/HistoryTree/HistoryTree.svelte'
 
   const { Story } = defineMeta({
     title: 'Components/Menu/Chat/History',
