+++ fai-rag-app/fai-sveltekit-frontend/src/lib/components/Menu/Chat/HistoryTree/historyUtils.ts
